requirements.txt
fastapi
uvicorn
vosk
openai
python-dotenv
pyttsx3
soundfile
librosa
scipy
sqlalchemy
asyncpg
aiofiles
httpx
pandas
celery
redis
black
flake8
pytest
gunicorn
prometheus_client
fastapi[all]
